10_AI_Workflow_App/ScriptSprout/backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "scriptsprout"
version = "0.1.0"
description = "ScriptSprout API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.0",
  "bcrypt>=4.0.0",
  "openai>=2.29.0",
  "pillow>=12.1.1",
  "chromadb>=1.5.5",
]

[dependency-groups]
dev = [
  "ruff>=0.8.0",
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "S", "C901"]
ignore = ["S105", "S107"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # allow assert in tests
